requirements.txt
torch==2.5.1
transformers==4.45.0
peft==0.11.1
xgboost>=2.0.0
datasets>=2.14.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
pyyaml>=6.0
requests>=2.31.0
biopython>=1.81
